refactor(api): :recycle: refactor of folder structure
