+++ apps/api/src/app.module.ts
@@ -2,17 +2,13 @@ import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ConfigModule } from '@nestjs/config';
-import { AuthModule } from './auth/auth.module';
-import { UserModule } from './user/user.module';
-import { AddressModule } from './address/address.module';
+import { moduleComponents } from './components';
 @Module({
   imports: [
     ConfigModule.forRoot({ isGlobal: true }),
     HttpModule,
-    AuthModule,
-    UserModule,
-    AddressModule,
     MongooseModule.forRoot(process.env.DB_URI!),
+    ...moduleComponents,
   ],
   controllers: [],
 })
