+++ apps/api/src/components/address/index.ts
@@ -1,0 +1,1 @@
+export * from './address.module';
