+++ apps/api/src/components/address/schema/address.model.ts
@@ -1,6 +1,6 @@
 import { Schema, SchemaFactory } from '@nestjs/mongoose';
 import { createAddressInput } from '@packages/models';
-import { zodToClass } from '../../lib/zod-to-schema';
+import { zodToClass } from '../../../lib/zod-to-schema';
 
 @Schema({
   timestamps: true,
