+++ apps/api/src/components/index.ts
@@ -1,0 +1,5 @@
+import { AddressModule } from './address';
+import { AuthModule } from './auth';
+import { UserModule } from './user';
+
+export const moduleComponents = [AddressModule, AuthModule, UserModule];
