+++ apps/api/src/components/user/user.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Get, Post, UsePipes } from '@nestjs/common';
 import { UserService } from './user.service';
-import { ZodValidationPipe } from '../pipes';
+import { ZodValidationPipe } from '../../core/pipes';
 import { createUserInput, TCreateUserInput } from '@packages/models';
 
 @Controller('user')
